Tkinter Projects/sales_management/requirements.txt
psycopg2-binary==2.9.1
pygame==2.0.1
